Remove schemas.ts and update claude.md
